feat(xmlio): make XmlReader implement AutoClosable
In a lots of usages, the XmlReader is created for a single stream.
It makes sense to use the reader within a try-with-resources block,
but this requires to implement AutoClosable.

Please note: the stream will not be closed by the underlying StAX parser.
But as the stream is usually within the same try-with-resources block,
it will be closed, too.