refactor!(xmlio): rename XmlWriter and remove interfaces
The XMLIO library is very basic and is mostly used to offer
writing capabilities to xoai-common and reading capabilities to
all other XOAI modules.

There are classes using the writer directly due to the name conflict
of io.gdcc.xoai.xml.XmlWriter and former io.gdcc.xoai.xmlio.XmlWriter.
With the renaming, this cannot happen anymore. This was also
annotated by Sonarcloud, and yes, it led to wrong practice.

The interfaces io.gdcc.xoai.xmlio.XmlWritable and
io.gdcc.xoai.xmlio.XmlReadable are not used anywhere and are duplicated
with the ones with the same name in io.gdcc.xoai.xml from xoai-common
module. They have been removed accordingly.