refactor(common): make EchoElement capable of writing from InputStream
Before, the XML string sent to the EchoElement was stuffed into a XmlReader
with a ByteArrayInputStream. Now we extend this to be capable of reading
from an arbitrary InputStream when given on object creation.

This commit also adds benchmarking tests using JMH to learn about the speed
decrease that parsing the XML from the input stream causes. It get's compared
to the "native" copy of input to output as seen in Dataverse. Run it via
`mvn -Pbenchmark clean test`

Another small change happened, too: instead of the deprecated Stack class
we now use the replacement "Deque".