feat(common): add CopyElement to paste XML stream data
This is a fast alternative to EchoElement, which does not do
any XML parsing before it copies XML data from an InputStream
into the XmlWriter.

It takes care of tricking the writer into accepting the data
without further addo, but REQUIRES that the writer already
contains a wrapping element (you cannot write at root with this).