refactor(common): change Metadata to use EchoElement, CopyElement or XOAIMetadata
All of these share the common interface XmlWritable, so we can store the element to write out
as metadata as this type (no switch or if necessary).

This also moves the data handling to using classes - they create the data and the XML writables,
this class is just used for the modeling of a fluent API.

It is linked to the creation of items, which are generated by the application using this library
via the repository interfaces. The application may decide how to fill in metadata when creating items.