refactor(data): specify return types in ItemRepository and helper #21
Instead of an Item, make the methods sending an identifier only, return an object of type
ItemIdentifier to make it more clear this does not carry metadata - this interface does not expose
a getMetadata(), which is done with the Item interface only.

The getItem() method is changed to send along a metadata format, so the application
can expose pregenerated or cached metadata within the specific format.

See also classes CopyElement and Metadata.copyFromStream()